Begin consolidating CDN references to CdnConstants.js
