+++ src/components/modals/CountryListModal.js
@@ -5,7 +5,8 @@ import { type AirshipBridge } from 'react-native-airship'
 import FastImage from 'react-native-fast-image'
 import { getCountry } from 'react-native-localize'
 
-import { COUNTRY_CODES, FLAG_LOGO_URL } from '../../constants/CountryConstants.js'
+import { FLAG_LOGO_URL } from '../../constants/CdnConstants'
+import { COUNTRY_CODES } from '../../constants/CountryConstants.js'
 import s from '../../locales/strings.js'
 import type { CountryData } from '../../types/types'
 import { type Theme, cacheStyles, useTheme } from '../services/ThemeContext.js'
